вынос в const, оформление шаблонов
